requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langsmith>=0.1.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
openai>=1.50.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
sentence-transformers>=3.0.0
hf_transfer>=0.1.6
pydantic>=2.0.0
pypdf>=5.0.0
langchain-tavily>=0.1.0 
tavily-python>=0.3.0
pdfminer.six>=20240706
pdfplumber>=0.11.4
pytesseract>=0.3.10
Pillow>=10.4.0
faiss-cpu>=1.8.0
reportlab>=4.2.0
Markdown>=3.7
weasyprint>=62.3